utilities/requirements.txt
argon2-cffi==23.1.0

